pyproject.toml
[project]
name = "TwitchViewerBOT"
version = "1.2"
description = "Twitch Bot for making fake viewers on your live"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests==2.31.0",
    "streamlink==5.5.1",
    "fake_useragent==1.1.3",
    "customtkinter==5.1.3",
    "pycryptodome==3.19.1",
    "cffi==1.15.1",
    "rich==13.7",
    "flask==3.0.3",
    "python-dotenv==1.0.1",
]

[project.optional-dependencies]
test = [
    "pytest==7.3.1"
]
